+++ src/kodo_python/encoder.hpp
@@ -54,6 +54,40 @@ pybind11::handle encoder_write_payload(Encoder& encoder)
     return PyByteArray_FromStringAndSize((char*)payload.data(), length);
 }
 
+template<class Encoder>
+pybind11::handle encoder_write_symbol(Encoder& encoder, pybind11::handle handle)
+{
+    PyObject* coefficients = handle.ptr();
+    assert(PyByteArray_Check(coefficients) && "The coefficients should be a "
+           "Python bytearray object");
+
+    std::vector<uint8_t> symbol(encoder.symbol_size());
+    uint32_t length = encoder.write_symbol(
+        symbol.data(), (uint8_t*)PyByteArray_AsString(coefficients));
+
+    return PyByteArray_FromStringAndSize((char*)symbol.data(), length);
+}
+
+
+template<class Encoder>
+pybind11::handle encoder_write_uncoded_symbol(Encoder& encoder, uint32_t index)
+{
+    std::vector<uint8_t> symbol(encoder.symbol_size());
+    uint32_t length = encoder.write_uncoded_symbol(symbol.data(), index);
+
+    return PyByteArray_FromStringAndSize((char*)symbol.data(), length);
+}
+
+template<class Encoder>
+pybind11::handle encoder_generate(Encoder& encoder)
+{
+    std::vector<uint8_t> coefficients(encoder.coefficient_vector_size());
+    encoder.generate(coefficients.data());
+
+    return PyByteArray_FromStringAndSize(
+        (char*)coefficients.data(), coefficients.size());
+}
+
 template<class Coder>
 struct extra_encoder_methods
 {
@@ -81,6 +115,20 @@ void encoder(pybind11::module& m, const std::string& name)
         .def("write_payload", &encoder_write_payload<encoder_type>,
              "Generate an encoded payload.\n\n"
              "\t:returns: The bytearray containing the encoded payload.\n")
+        .def("write_symbol", &encoder_write_symbol<encoder_type>,
+             arg("coefficients"),
+             "Generate an encoded symbol using the given coefficients.\n\n"
+             "\t:param coefficients: The coding coefficients.\n"
+             "\t:returns: The bytearray containing the encoded symbol.\n")
+        .def("write_uncoded_symbol",
+             &encoder_write_uncoded_symbol<encoder_type>,
+             arg("index"),
+             "Return a systematic symbol for the given symbol index.\n\n"
+             "\t:param index: The symbol index with the coding block.\n"
+             "\t:returns: The bytearray containing the systematic symbol.\n")
+        .def("generate", &encoder_generate<encoder_type>,
+             "Generate some coding coefficients.\n\n"
+             "\t:returns: The bytearray containing the coding coefficients.\n")
         .def("set_const_symbols", &set_const_symbols<encoder_type>,
              arg("symbols"),
              "Set the symbols to be encoded.\n\n"